+++ src/main/java/fr/estecka/variantscit/EquippableCache.java
@@ -12,7 +12,7 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.equipment.EquipmentAssets;
 import net.minecraft.world.item.equipment.Equippable;
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
-import fr.estecka.variantscit.reload.EModuleContext;
+import fr.estecka.variantscit.reload.EModuleHook;
 
 public class EquippableCache
 {
@@ -60,7 +60,7 @@ public class EquippableCache
 		}
 
 		VariantsCitMod.LOGGER.PushLabel(stack.getItem());
-		ResourceLocation assetId = VariantsCitMod.GetModules().GetModelForItem(EModuleContext.EQUIPPABLE, stack);
+		ResourceLocation assetId = VariantsCitMod.GetModules().GetModelForItem(EModuleHook.EQUIPPABLE, stack);
 		VariantsCitMod.LOGGER.PopLabel();
 
 		if (assetId == null)