+++ src/main/java/fr/estecka/variantscit/mixin/ItemModelManagerMixin.java
@@ -6,7 +6,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import fr.estecka.variantscit.VariantsCitMod;
-import fr.estecka.variantscit.reload.EModuleContext;
+import fr.estecka.variantscit.reload.EModuleHook;
 import net.minecraft.client.renderer.item.ItemModelResolver;
 import net.minecraft.core.component.DataComponentType;
 import net.minecraft.resources.ResourceLocation;
@@ -24,7 +24,7 @@ public class ItemModelManagerMixin
 		ResourceLocation modelId = null;
 		
 		VariantsCitMod.LOGGER.PushLabel(stack.getItem());
-		modelId = VariantsCitMod.GetModules().GetModelForItem(EModuleContext.ITEM_MODEL, stack);
+		modelId = VariantsCitMod.GetModules().GetModelForItem(EModuleHook.ITEM_MODEL, stack);
 		VariantsCitMod.LOGGER.PopLabel();
 
 		if (modelId == null)